src/read/requirements.txt
git+https://github.com/nashspence/home-index.git
debugpy
easyocr
numpy
pdf2image
pillow
python-magic
rawpy
wand